+++ packages/stream_video_flutter/test/src/call_participants/call_participants_order_test.dart
@@ -15,6 +15,7 @@ import '../mocks.dart';
 void main() {
   setUp(() {
     VisibilityDetectorController.instance.updateInterval = Duration.zero;
+    ViewportVisibilityRegistry.defaultReleaseGrace = Duration.zero;
   });
 
   CallParticipantState participant(